# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Education Consultancy Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List agents
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentsListResponse"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentBody"
      responses:
        "201":
          description: Created agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get agent by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentBody"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: agentId
          in: query
          schema:
            type: integer
        - name: stage
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationsListResponse"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationBody"
      responses:
        "201":
          description: Created application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationBody"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    delete:
      operationId: deleteApplication
      tags: [applications]
      summary: Delete application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /applications/{id}/notes:
    get:
      operationId: getApplicationNotes
      tags: [applications]
      summary: Get application notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: addApplicationNote
      tags: [applications]
      summary: Add note to application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

  # DOCUMENTS
...
  /blog:
    get:
      operationId: listBlogPosts
      tags: [content]
      summary: List blog posts
      parameters:
        - name: locale
          in: query
          schema:
            type: string
        - name: published
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPostsListResponse"
    post:
      operationId: createBlogPost
      tags: [content]
      summary: Create blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPostBody"
      responses:
        "201":
          description: Created blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [content]
      summary: Get blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: locale
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBlogPost
      tags: [content]
      summary: Update blog post
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBlogPostBody"
      responses:
        "200":
          description: Updated blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [content]
      summary: Delete blog post
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [content]
      summary: List announcements
      parameters:
        - name: audience
          in: query
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [content]
      summary: Create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementBody"
      responses:
        "201":
          description: Created announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [content]
      summary: Update announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnnouncementBody"
      responses:
        "200":
          description: Updated announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [content]
      summary: Delete announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SETTINGS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - total
        - page
        - limit
        - totalPages
...
    Note:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        authorId:
          type: integer
        authorName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - authorId
        - createdAt
    CreateNoteBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...
    Agent:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        commissionRate:
          type: ["number", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - status
        - createdAt
        - updatedAt
    CreateAgentBody:
      type: object
      properties:
        userId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        commissionRate:
          type: ["number", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - firstName
        - lastName
        - status
    UpdateAgentBody:
      type: object
      properties:
        companyName:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        commissionRate:
          type: ["number", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    AgentsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    Application:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        programId:
          type: ["integer", "null"]
        universityId:
          type: ["integer", "null"]
        agentId:
          type: ["integer", "null"]
        stage:
          type: string
        intake:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        programName:
          type: ["string", "null"]
        universityName:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        tuitionFee:
          type: ["number", "null"]
        scholarship:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - stage
        - createdAt
        - updatedAt
    CreateApplicationBody:
      type: object
      properties:
        studentId:
          type: integer
        programId:
          type: ["integer", "null"]
        universityId:
          type: ["integer", "null"]
        agentId:
          type: ["integer", "null"]
        stage:
          type: string
        intake:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        programName:
          type: ["string", "null"]
        universityName:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        tuitionFee:
          type: ["number", "null"]
        scholarship:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - studentId
        - stage
    UpdateApplicationBody:
      type: object
      properties:
        programId:
          type: ["integer", "null"]
        universityId:
          type: ["integer", "null"]
        agentId:
          type: ["integer", "null"]
        stage:
          type: ["string", "null"]
        intake:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        programName:
          type: ["string", "null"]
        universityName:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        tuitionFee:
          type: ["number", "null"]
        scholarship:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    ApplicationsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
...
    BlogPost:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        locale:
          type: string
        published:
          type: boolean
        authorId:
          type: ["integer", "null"]
        featuredImageUrl:
          type: ["string", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - title
        - locale
        - published
        - createdAt
        - updatedAt
    CreateBlogPostBody:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        locale:
          type: string
        published:
          type: boolean
        featuredImageUrl:
          type: ["string", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
      required:
        - slug
        - title
        - locale
        - published
    UpdateBlogPostBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        locale:
          type: ["string", "null"]
        published:
          type: ["boolean", "null"]
        featuredImageUrl:
          type: ["string", "null"]
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
    BlogPostsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        audience:
          type: string
        isActive:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - audience
        - isActive
        - createdAt
    CreateAnnouncementBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        audience:
          type: string
        isActive:
          type: boolean
        expiresAt:
          type: ["string", "null"]
      required:
        - title
        - content
        - audience
        - isActive
    UpdateAnnouncementBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        audience:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
        expiresAt:
          type: ["string", "null"]
...