# lib/api-zod/src/generated/types/createApplicationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Education Consultancy Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateApplicationBody {
  studentId: number;
  /** @nullable */
  programId?: number | null;
  /** @nullable */
  universityId?: number | null;
  /** @nullable */
  agentId?: number | null;
  stage: string;
  /** @nullable */
  intake?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  programName?: string | null;
  /** @nullable */
  universityName?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  tuitionFee?: number | null;
  /** @nullable */
  scholarship?: number | null;
  /** @nullable */
  notes?: string | null;
}